re-organized venv-selector
